// src/common/enums/order.enum.ts
export enum OrderStatusEnum{
    Pending = 0,
    Placed = 1,
    OnWay = 2,
    Delivered = 3,
    Cancel = 4,
}
export enum OrderStatusNameEnum{
    Pending = "Pending",
    Placed = "Placed",
    OnWay = "On_Way",
    Delivered = "Delivered",
    Cancel = "Cancel",
}


export enum PaymentEnum{
    Card = "Card",
    Cash="Cash",
}